+++ src/components/CampaignDetails/CampaignDetails.tsx
@@ -20,6 +20,7 @@ import useCreateCampaignContext from 'hooks/useCreateCampaignContext'
 import useCustomNotifications from 'hooks/useCustomNotifications'
 import { CustomConfirmModal } from 'components/common/Modals'
 import CatsLocsFormatted from './CatsLocsFormatted'
+import { AdminActions } from './AdminActions'
 
 const useStyles = createStyles((theme) => ({
   wrapper: {
@@ -53,7 +54,7 @@ const useStyles = createStyles((theme) => ({
   }
 }))
 
-const CampaignDetails = () => {
+const CampaignDetails = ({ isAdminPanel }: { isAdminPanel?: boolean }) => {
   const { classes, cx } = useStyles()
   const { campaignsData, updateCampaignDataById, changeCampaignStatus } = useCampaignsData()
   const { updateCampaignFromDraft } = useCreateCampaignContext()
@@ -113,7 +114,7 @@ const CampaignDetails = () => {
         <>
           <Container fluid className={classes.wrapper}>
             <Grid>
-              <Grid.Col span={6}>
+              <Grid.Col md={12} xl={6}>
                 <Text weight="bold" size="sm" pb="sm" className={classes.lighterColor}>
                   Overview
                 </Text>
@@ -241,7 +242,7 @@ const CampaignDetails = () => {
                   />
                 </div>
               </Grid.Col>
-              <Grid.Col span={6}>
+              <Grid.Col md={12} xl={6}>
                 <Grid>
                   <Grid.Col span={12}>
                     <Text weight="bold" size="sm" pb="sm" className={classes.lighterColor}>
@@ -283,7 +284,13 @@ const CampaignDetails = () => {
                             lineHeight="sm"
                             textSize="sm"
                             title={`${item.banner?.format.w}x${item.banner?.format.h}`}
-                            value={<MediaThumb adUnit={item} previewOnClick />}
+                            value={
+                              <MediaThumb
+                                adUnit={item}
+                                previewOnClick
+                                title={`Target URL: ${item.banner?.targetUrl}`}
+                              />
+                            }
                             noBorder={isLast}
                           />
                         )
@@ -347,6 +354,14 @@ const CampaignDetails = () => {
                   </Grid.Col>
                 </Grid>
               </Grid.Col>
+
+              {isAdminPanel && campaign.status === CampaignStatus.inReview && (
+                <Grid>
+                  <Grid.Col>
+                    <AdminActions item={campaign} />
+                  </Grid.Col>
+                </Grid>
+              )}
             </Grid>
           </Container>
           <CustomConfirmModal
