+++ src/components/common/Modals/index.ts
@@ -1,6 +1,5 @@
 import CreativePreviewModal from './CreativePreviewModal/CreativePreviewModal'
 import CampaignPeriodModal from './CampaignPeriodModal/CampaignPeriodModal'
-import AdminCampaignModal from './AdminCampaignModal/AdminCampaignModal'
 import LaunchCampaignModal from './LaunchCampaignModal/LaunchCampaignModal'
 import SuccessModal from './SuccessModal'
 import CustomConfirmModal from './CustomConfirmModal/CustomConfirmModal'
@@ -8,7 +7,6 @@ import CustomConfirmModal from './CustomConfirmModal/CustomConfirmModal'
 export {
   CreativePreviewModal,
   CampaignPeriodModal,
-  AdminCampaignModal,
   LaunchCampaignModal,
   SuccessModal,
   CustomConfirmModal
